Split tracing types out of Types module
The types module can be simpler and clearer this way. Quicker to get a
grasp of the key types.

Also export a few more things via the top level Network.Mux so most
users of the library don't need to import so many modules.
