+++ network-mux/network-mux.cabal
@@ -54,6 +54,7 @@ library
                        Network.Mux.Interface
                        Network.Mux.Time
                        Network.Mux.Types
+                       Network.Mux.Trace
                        Network.Mux.Bearer.Pipe
                        Network.Mux.Bearer.Queues
                        Network.Mux.Bearer.Socket
@@ -75,6 +76,7 @@ test-suite test-network-mux
                        Network.Mux.Interface
                        Network.Mux.Time
                        Network.Mux.Types
+                       Network.Mux.Trace
                        Network.Mux.Bearer.Pipe
                        Network.Mux.Bearer.Queues
                        Network.Mux.Bearer.Socket
