+++ network-mux/src/Network/Mux/Bearer/Queues.hs
@@ -17,7 +17,6 @@ import           Control.Monad.Class.MonadThrow
 import           Control.Tracer
 
 import qualified Network.Mux as Mx
-import qualified Network.Mux.Interface as Mx
 import           Network.Mux.Types (MuxBearer)
 import qualified Network.Mux.Types as Mx
 import qualified Network.Mux.Codec as Mx
