+++ network-mux/src/Network/Mux/Bearer/Socket.hs
@@ -22,6 +22,7 @@ import qualified Network.Socket.ByteString.Lazy as Socket (recv, sendAll)
 import qualified Network.Mux as Mx
 import           Network.Mux.Types (MuxBearer)
 import qualified Network.Mux.Types as Mx
+import qualified Network.Mux.Trace as Mx
 import qualified Network.Mux.Codec as Mx
 import qualified Network.Mux.Time as Mx
